+++ src/common/fileUtil.ts
@@ -1,8 +1,9 @@
 import { existsSync, mkdirSync } from 'fs';
 import { dirname, parse } from 'path';
 import * as vscode from 'vscode';
+import { Global } from './global';
 
-export function wrieteFile(path: string, buffer: Buffer) {
+export function writeFile(path: string, buffer: Buffer) {
     const dir = dirname(path)
     if (!existsSync(dir)) {
         mkdirSync(dir, { recursive: true })
@@ -11,7 +12,7 @@ export function wrieteFile(path: string, buffer: Buffer) {
 }
 
 export function adjustImgPath(uri: vscode.Uri, withworkspace: boolean = false) {
-    let imgPath = vscode.workspace.getConfiguration("vscode-office").get<string>("pasterImgPath")
+    const imgPath = Global.getConfig<string>("pasterImgPath")
         .replace("${fileName}", parse(uri.fsPath).name.replace(/\s/g, ''))
         .replace("${now}", new Date().getTime() + "")
     return {
@@ -29,7 +30,7 @@ export function adjustImgPath(uri: vscode.Uri, withworkspace: boolean = false) {
 export function getWorkspacePath(uri: vscode.Uri): string {
     const folders = vscode.workspace.workspaceFolders;
     if (!folders || folders.length == 0) return '';
-    let workspacePath = folders[0]?.uri?.fsPath;
+    const workspacePath = folders[0]?.uri?.fsPath;
     if (folders.length > 1) {
         for (const folder of folders) {
             if (uri.fsPath.includes(folder.uri.fsPath)) {