+++ src/provider/markdownEditorProvider.ts
@@ -1,11 +1,13 @@
-import { adjustImgPath, getWorkspacePath, wrieteFile } from '@/common/fileUtil';
+import { adjustImgPath, getWorkspacePath, writeFile } from '@/common/fileUtil';
 import { readFileSync } from 'fs';
 import { basename, isAbsolute, parse, resolve } from 'path';
 import * as vscode from 'vscode';
-import { Hanlder } from '../common/handler';
+import { Handler } from '../common/handler';
 import { Util } from '../common/util';
 import { Holder } from '../service/markdown/holder';
 import { MarkdownService } from '../service/markdownService';
+import { Global } from '@/common/global';
+import { platform } from 'os';
 
 /**
  * support view and edit office files.
@@ -24,7 +26,7 @@ export class MarkdownEditorProvider implements vscode.CustomTextEditorProvider {
 
     private getFolders(): vscode.Uri[] {
         const data = [];
-        for (var i = 65; i <= 90; i++) {
+        for (let i = 65; i <= 90; i++) {
             data.push(vscode.Uri.file(`${String.fromCharCode(i)}:/`))
         }
         return data;
@@ -38,12 +40,12 @@ export class MarkdownEditorProvider implements vscode.CustomTextEditorProvider {
             enableScripts: true,
             localResourceRoots: [vscode.Uri.file("/"), ...this.getFolders()]
         }
-        const handler = Hanlder.bind(webviewPanel, uri);
+        const handler = Handler.bind(webviewPanel, uri);
         this.handleMarkdown(document, handler, folderPath)
         handler.on('developerTool', () => vscode.commands.executeCommand('workbench.action.toggleDevTools'))
     }
 
-    private handleMarkdown(document: vscode.TextDocument, handler: Hanlder, folderPath: vscode.Uri) {
+    private handleMarkdown(document: vscode.TextDocument, handler: Handler, folderPath: vscode.Uri) {
 
         const uri = document.uri;
         const webview = handler.panel.webview;
@@ -95,7 +97,7 @@ export class MarkdownEditorProvider implements vscode.CustomTextEditorProvider {
         }).on("img", (img) => {
             const { relPath, fullPath } = adjustImgPath(uri)
             const imagePath = isAbsolute(fullPath) ? fullPath : `${resolve(uri.fsPath, "..")}/${relPath}`.replace(/\\/g, "/");
-            wrieteFile(imagePath, Buffer.from(img, 'binary'))
+            writeFile(imagePath, Buffer.from(img, 'binary'))
             const fileName = parse(relPath).name;
             vscode.env.clipboard.writeText(`![${fileName}](${relPath})`)
             vscode.commands.executeCommand("editor.action.clipboardPasteAction")
@@ -112,21 +114,35 @@ export class MarkdownEditorProvider implements vscode.CustomTextEditorProvider {
         }).on("export", (option) => {
             vscode.commands.executeCommand('workbench.action.files.save');
             new MarkdownService(this.context).exportMarkdown(uri, option)
-        }).on("theme", () => {
-            vscode.commands.executeCommand('workbench.action.selectTheme');
+        }).on("theme", async (theme) => {
+            if (!theme) {
+                const themes = ["Auto", "|", "Light", "Solarized", "|", "Dracula", "Github Dark"]
+                const editorTheme = Global.getConfig('editorTheme');
+                const themeItems: vscode.QuickPickItem[] = themes.map(theme => {
+                    if (theme == '|') return { label: '|', kind: vscode.QuickPickItemKind.Separator }
+                    return { label: theme, description: theme == editorTheme ? 'Current' : undefined }
+                })
+                theme = await vscode.window.showQuickPick(themeItems, { placeHolder: "Select Editor Theme" });
+                if (!theme) return
+            }
+            handler.emit('theme', theme.label)
+            Global.updateConfig('editorTheme', theme.label)
         }).on("saveOutline", (enable) => {
             config.update("openOutline", enable, true)
         }).on('developerTool', () => {
             vscode.commands.executeCommand('workbench.action.toggleDevTools')
         })
 
-        const basePath = vscode.workspace.getConfiguration("vscode-office").get<boolean>("workspacePathAsImageBasePath") ?
+        const basePath = Global.getConfig('workspacePathAsImageBasePath') ?
             vscode.Uri.file(getWorkspacePath(folderPath)) : folderPath;
         const baseUrl = webview.asWebviewUri(basePath).toString().replace(/\?.+$/, '').replace('https://git', 'https://file');
         webview.html = Util.buildPath(
             readFileSync(`${this.extensionPath}/resource/vditor/index.html`, 'utf8')
                 .replace("{{rootPath}}", rootPath)
-                .replace("{{baseUrl}}", baseUrl),
+                .replace("{{baseUrl}}", baseUrl)
+                .replace(`{{configs}}`, JSON.stringify({
+                    platform: platform()
+                })),
             webview, contextPath);
     }