+++ src/provider/officeViewerProvider.ts
@@ -1,16 +1,15 @@
-import { ZipService } from '@/service/zip/zipService';
-import axios from 'axios';
+import { ReactApp } from '@/common/reactApp';
 import { spawn } from 'child_process';
-import { existsSync, mkdirSync, readdirSync, readFileSync } from 'fs';
+import { existsSync, mkdirSync, readFileSync } from 'fs';
 import { tmpdir } from 'os';
-import { basename, extname, parse, resolve } from 'path';
+import { extname, parse } from 'path';
 import { TextEncoder } from 'util';
 import * as vscode from 'vscode';
-import { workspace } from 'vscode';
-import { Hanlder } from '../common/handler';
 import { Output } from '../common/Output';
+import { Handler } from '../common/handler';
 import { Util } from '../common/util';
-import { ViewManager } from '@/common/viewManager';
+import { handleImage, isImage } from './handlers/imageHanlder';
+import { handleZip } from './handlers/zipHandler';
 
 /**
  * support view office files
@@ -21,6 +20,11 @@ export class OfficeViewerProvider implements vscode.CustomReadonlyEditorProvider
 
     constructor(private context: vscode.ExtensionContext) {
         this.extensionPath = context.extensionPath;
+    }
+
+    bindCustomEditors(viewOption: { webviewOptions: vscode.WebviewPanelOptions }) {
+        const viewers = ['cweijan.officeViewer', 'cweijan.imageViewer', 'cweijan.htmlViewer', 'cweijan.classViewer']
+        return viewers.map(viewer => vscode.window.registerCustomEditorProvider(viewer, this, viewOption))
     }
 
     public openCustomDocument(uri: vscode.Uri, openContext: vscode.CustomDocumentOpenContext, token: vscode.CancellationToken): vscode.CustomDocument | Thenable<vscode.CustomDocument> {
@@ -46,19 +50,19 @@ export class OfficeViewerProvider implements vscode.CustomReadonlyEditorProvider
             })
         }
 
-        const handler = Hanlder.bind(webviewPanel, uri);
+        let route: string;
+        const handler = Handler.bind(webviewPanel, uri);
         handler
+            .on("editInVSCode", (full: boolean) => {
+                const side = full ? vscode.ViewColumn.Active : vscode.ViewColumn.Beside;
+                vscode.commands.executeCommand('vscode.openWith', uri, "default", side);
+            })
             .on('developerTool', () => vscode.commands.executeCommand('workbench.action.toggleDevTools'))
             .on("init", send)
 
-        if (ext.match(/\.(jpg|png|svg|gif|apng|bmp|ico|cur|jpeg|pjpeg|pjp|tif|webp)$/i)) {
-            this.handleImage(uri, webview)
-            handler.on("fileChange", () => {
-                this.handleImage(uri, webview)
-            }).on('developerTool', () => {
-                vscode.commands.executeCommand('workbench.action.toggleDevTools')
-            })
-            return;
+        if (isImage(ext)) {
+            handleImage(handler, uri, webview)
+            return ReactApp.view(webview, { route: 'image' })
         }
 
         switch (ext) {
@@ -67,18 +71,20 @@ export class OfficeViewerProvider implements vscode.CustomReadonlyEditorProvider
             case ".xls":
             case ".csv":
             case ".ods":
-                htmlPath = this.handleXlsx(uri, handler)
+                route = 'excel';
+                handler.on("fileChange", send)
                 break;
             case ".docx":
             case ".dotx":
-                htmlPath = 'word.html'
+                route = 'word'
                 handler.on("fileChange", send)
                 break;
             case ".jar":
             case ".zip":
             case ".apk":
             case ".vsix":
-                this.handleZip(webview, uri, handler);
+                route = 'zip';
+                handleZip(uri, handler);
                 break;
             case ".pdf":
                 this.handlePdf(webview);
@@ -86,6 +92,7 @@ export class OfficeViewerProvider implements vscode.CustomReadonlyEditorProvider
                 break;
             case ".ttf":
             case ".woff":
+            case ".woff2":
             case ".otf":
                 this.handleFont(handler)
                 break;
@@ -102,81 +109,31 @@ export class OfficeViewerProvider implements vscode.CustomReadonlyEditorProvider
             default:
                 vscode.commands.executeCommand('vscode.openWith', uri, "default");
         }
+        if (route) return ReactApp.view(webview, { route })
 
         if (htmlPath != null) {
             webview.html = Util.buildPath(readFileSync(this.extensionPath + "/resource/" + htmlPath, 'utf8'), webview, this.extensionPath + "/resource")
-                .replace("$autoTheme", workspace.getConfiguration("vscode-office").get<boolean>("autoTheme") + '')
         }
 
     }
-
-    async handleZip(webview: vscode.Webview, uri: vscode.Uri, handler: Hanlder) {
-        let data = await ViewManager.readContent()
-        data = await ViewManager.buildPath(data, webview);
-        webview.html = data
-        new ZipService(uri, handler).bind();
-    }
-
-    private handleImage(uri: vscode.Uri, webview: vscode.Webview) {
-
-        const folderPath = vscode.Uri.file(resolve(uri.fsPath, ".."));
-        const files = readdirSync(folderPath.fsPath)
-        let text = "";
-        let current = 0;
-        let i = 0;
-        const currentFile = basename(uri.fsPath)
-        if (uri.scheme == 'git') {
-            const href = webview.asWebviewUri(uri);
-            text += `<a href="${href}" title="${basename(uri.fsPath)}"> <img src="${href}" > </a>`
-        } else {
-            for (const file of files) {
-                if (currentFile == file) {
-                    current = i;
-                }
-                if (file.match(/\.(jpg|png|svg|gif|apng|bmp|ico|cur|jpeg|pjpeg|pjp|tif|tiff|webp)$/i)) {
-                    i++;
-                    const resUri = vscode.Uri.file(folderPath.fsPath + "/" + file);
-                    const resource = webview.asWebviewUri(resUri).with({ query: `nonce=${Date.now().toString()}` }).toString();
-                    text += `<a href="${resource}" title="${file}"> <img src="${resource}" > </a>`
-                }
-            }
-        }
-
-        webview.html =
-            Util.buildPath(readFileSync(this.extensionPath + "/resource/lightgallery/lg.html", 'utf8'), webview, this.extensionPath + "/resource/lightgallery")
-                .replace("{{content}}", text).replace("{{current}}", `${current}`);
-    }
-
 
     private handlePdf(webview: vscode.Webview) {
-        const baseUrl = webview.asWebviewUri(vscode.Uri.file(this.extensionPath + "/resource/pdf"))
-            .toString().replace(/\?.+$/, '').replace('https://git', 'https://file');
+        const baseUrl = this.getBaseUrl(webview, 'pdf')
         webview.html = readFileSync(this.extensionPath + "/resource/pdf/viewer.html", 'utf8').replace("{{baseUrl}}", baseUrl)
     }
 
-    private handleFont(handler: Hanlder) {
+    private handleFont(handler: Handler) {
         const webview = handler.panel.webview;
-        webview.html = Util.buildPath(
-            readFileSync(`${this.extensionPath}/resource/font/index.html`, 'utf8'),
-            webview, `${this.extensionPath}/resource/font`
-        )
+        const baseUrl = this.getBaseUrl(webview, 'font')
+        webview.html = readFileSync(`${this.extensionPath}/resource/font/index.html`, 'utf8')
+            .replace('{{baseUrl}}', baseUrl)
     }
 
-
-    private handleXlsx(uri: vscode.Uri, handler: Hanlder) {
-        const enc = new TextEncoder();
-        handler.on("save", async (content) => {
-            Util.confirm(`Save confirm`, 'Are you sure you want to save? this will lose all formatting.', async () => {
-                await vscode.workspace.fs.writeFile(uri, new Uint8Array(content))
-                handler.emit("saveDone")
-            })
-        }).on("saveCsv", async (content) => {
-            await vscode.workspace.fs.writeFile(uri, enc.encode(content))
-            handler.emit("saveDone")
-        })
-        return "excel.html"
+    private getBaseUrl(webview: vscode.Webview, path: string) {
+        const baseUrl = webview.asWebviewUri(vscode.Uri.file(`${this.extensionPath}/resource/${path}`))
+            .toString().replace(/\?.+$/, '').replace('https://git', 'https://file')
+        return baseUrl;
     }
-
 
     private async handleClass(uri: vscode.Uri, panel: vscode.WebviewPanel) {
         if (uri.scheme != "file") {